refactor: splitting components code to each folder
