+++ src/Components/index.js
@@ -1,14 +1,10 @@
 import styled from 'styled-components';
 import { Link } from 'react-router-dom';
-import media from 'styled-media-query';
 
-const Button = styled.button`
-  background: #007afd;
-  padding: 0.25em 1em;
-  border: none;
-  border-radius: 0 8px 8px 0;
-  height: 35px;
-  width: 15%;
+export const StyledLink = styled(Link)`
+  color: #007afd;
+  font-size: 20px;
+  margin-bottom: 0.3rem;
 `;
 
 export const HeaderSection = styled.section`
@@ -20,39 +16,6 @@ export const H1 = styled.h1`
   font-size: 36px;
   font-weight: 700;
   margin: 0;
-`;
-
-export const Input = styled.input`
-  height: 35px;
-  width: 85%;
-  padding-left: 15px;
-  color: rgba(0, 0, 0, 0.54);
-  background: #ebebeb;
-  border: none;
-  border-radius: 8px 0 0 8px;
-  font-size: 16px;
-`;
-
-export const StyledLink = styled(Link)`
-  color: #007afd;
-  font-size: 20px;
-  margin-bottom: 0.3rem;
-`;
-
-// no decorate link
-export const NDLink = styled(Link)`
-  text-decoration: none;
-  color: black;
-`;
-
-export const CountryWrapper = styled.div`
-  padding: 7px 7px;
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-  align-items: center;
-  margin-bottom: 0.75rem;
-  border-bottom: 1px solid grey;
 `;
 
 export const FlexRow = styled.div`
@@ -90,211 +53,44 @@ export const Arrow = styled.div`
   }}
 `;
 
-export const ErrorWrapper = styled.div`
-  margin-top: 1rem;
-  width: 100%;
-  padding: 20px;
-  background-color: rgba(255, 0, 0, 0.6);
-  border-radius: 6px;
-  color: white;
-`;
-
-export const ErrorH3 = styled.h3`
-  margin: 0;
-  font-size: 18px;
-  font-weight: 500;
-  margin-bottom: 10px;
-`;
-
-export const ErrorP = styled.p`
-  margin: 0;
-  font-size: 14px;
-`;
-
-export const ErrorHeader = styled.div`
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-`;
-
-export const CloseIcon = styled.div`
-  position: relative;
-  width: 10px;
-  height: 10px;
-
-  &:before,
-  &:after {
-    position: absolute;
-    content: ' ';
-    height: 19px;
-    width: 2px;
-    background-color: #fff;
-  }
-  &:before {
-    transform: rotate(45deg);
-  }
-  &:after {
-    transform: rotate(-45deg);
-  }
-`;
-
-export const CountryPageWrapper = styled.div`
-  position: absolute;
-  top: 0;
-  left: 0;
-  width: 100%;
-  margin-bottom: 3rem;
-  background-color: #e5e5e5;
-
-  ${media.greaterThan('large')`
-    background-color: white;
-  `}
-`;
-
-export const CountryPageInner = styled.div`
-  ${media.greaterThan('large')`
-    width: 50%;
-    margin: 0 auto;
-  `}
-`;
-
-export const SectionWrapper = styled.div`
-  margin-bottom: 2rem;
-`;
-
-export const DataTableHeader = styled.h2`
-  font-size: 16px;
-  font-weight: 400;
-  color: rgb(0, 0, 0, 0.6);
-  text-transform: uppercase;
-  margin-left: 10px;
-`;
-
-export const Flag = styled.img`
-  margin-top: 1rem;
-  height: 180px;
-  display: block;
-  margin-left: auto;
-  margin-right: auto;
-  border: 1px solid grey;
-  border-radius: 10px;
-`;
-
-export const DataTable = styled.div`
-  background-color: white;
-  padding: 10px;
-  display: flex;
-  flex-direction: column;
-
-  ${media.greaterThan('large')`
-    border-radius: 10px;
-    border: 1px solid #f5f5f5;
-    box-shadow: rgb(0,0,0,0.1) 0px 2px 4px 0px
-  `}
-`;
-
-export const DataBox = styled.div`
-  padding: 10px 8px 10px 8px;
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-  border-bottom: 1px solid rgba(0, 0, 0, 0.1);
-
-  &:last-child {
-    border-bottom: none;
-  }
-`;
-
-export const DataName = styled.span`
-  text-transform: capitalize;
-  font-size: 14px;
-  font-weight: 700;
-`;
-
-export const DataValue = styled.span`
-  font-size: 14px;
-  font-weight: 500;
-  color: rgb(0, 0, 0, 0.5);
-`;
-
-export const CPHeader = styled.div`
-  display: flex;
-  background-color: white;
-  padding: 12px 10px;
-  border-bottom: 1px solid rgba(0, 0, 0, 0.1);
-
-  ${media.greaterThan('large')`
-  `}
-`;
-
-export const CPHeaderTitle = styled.div`
-  font-weight: 500;
-  font-size: 18px;
-`;
-
-export const BackSearchBtn = styled.button`
-  margin-right: auto;
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  color: #007afd;
+export const NDLink = styled(Link)`
   text-decoration: none;
-  background: white;
-  border: none;
-  cursor: pointer;
-`;
-export const BackSearchT = styled.span`
-  font-weight: 700;
-  font-size: 16px;
-  margin-left: 1px;
+  color: black;
 `;
 
-export const EmptyT = styled.div`
-  margin-left: auto;
-  visibility: hidden;
-`;
+// export const ErrorHeader = styled.div`
+//   display: flex;
+//   flex-direction: row;
+//   justify-content: space-between;
+// `;
 
-export const CurrencyDataBox = styled.div`
-  padding: 10px 8px 10px 8px;
-  display: flex;
-  flex-direction: column;
-`;
+// export const CloseIcon = styled.div`
+//   position: relative;
+//   width: 10px;
+//   height: 10px;
 
-export const ExchangeBox = styled.div`
-  padding: 10px 8px 10px 8px;
-  height: 80px;
-  display: flex;
-  flex-direction: row;
-  justify-content: space-around;
-`;
+//   &:before,
+//   &:after {
+//     position: absolute;
+//     content: ' ';
+//     height: 19px;
+//     width: 2px;
+//     background-color: #fff;
+//   }
+//   &:before {
+//     transform: rotate(45deg);
+//   }
+//   &:after {
+//     transform: rotate(-45deg);
+//   }
+// `;
 
-export const ExchangeInnerBox = styled.div`
-  height: 100%;
-  width: 150px;
-  border: 1px solid rgb(0, 0, 0, 0.1);
-  border-radius: 10px;
+export const CountryWrapper = styled.div`
+  padding: 7px 7px;
   display: flex;
   flex-direction: row;
-  justify-content: center;
+  justify-content: space-between;
   align-items: center;
-  font-size: 24px;
+  margin-bottom: 0.75rem;
+  border-bottom: 1px solid grey;
 `;
-
-export const Select = styled.select`
-  font-size: 14px;
-  width: 90%;
-  border: 1px #fff solid;
-  background-color: white;
-  padding: 0.6em 1.4em 0.5em 0.8em;
-  box-sizing: border-box;
-  margin: 0;
-
-  &:hover {
-    background-color: #f0f0f0;
-    cursor: pointer;
-  }
-`;
-
-export const Option = styled.option``;
-
-export default Button;
