+++ src/Search/index.js
@@ -1,7 +1,8 @@
 import React, { useState } from 'react';
 import axios from 'axios';
 import { useLocation, useHistory } from 'react-router-dom';
-import Button, { Input, H1, HeaderSection } from '../Components';
+import { H1, HeaderSection } from '../Components';
+import Button, { Input } from './Components';
 import ShowResults from './ShowResults';
 
 const Search = () => {
